refactor: adapt to Disman's control integration
With recent changes Disman learned to handle Control files reading and writing
on its own. All control data is transferred as part of the Config and Output
objects.

This allows us to remove all control handling in KDisplay and directly evaluate
instead the data sent by Disman.

This patch makes KDisplay first work again with the changed API and operational
flow of Disman. In a second step the additional functionality that Disman now
provides should be reaped. That means foremost the automatic control setting
for resolution and refresh rate.