feat(kcm): configure automatic resolution and refresh rate
Leverages new API functionality of Disman to set if a mode should be determined
automatically or if the user wants to set resolution and/or refresh rate
manually.

In case a property is set automatically the combo box is greyed out. Otherwise
we show an ascending list of available refresh rates for the currently set
resolution. In comparision to before this now includes all possible resolutions
independent on how small their delta is.