feat(kcm): move resolution setting down
With auto resolution being the default and normally all that is needed we do
not need to present the combo box at the top of the page. Scale and rotation
are more important so show these on top and move the resolution down above the
refresh rate setting with which together it also makes more sense.