fix(kded): update orientation on every change
With the new model of configuration storing in Disman the KDisplay daemon
does not always apply configurations anymore. Move the orientation udpate
so that it gets executed in any case. Also we start in the monitoring state.

This change is somewhat experimental and we need to check back if it works
without issues for devices with orientation sensor.