feat: disable KDisplay's config control
KDisplay's code paths for generating special configs for laptops with closed
lid or when in a docking station and when the user selects an action in the
OSD is currently not well tested and showed problems in some minimal testing.

Since this is rather superficial functioanlity disable it for now until we can
fully adapt it to the recent changes in Disman. It might then be also an option
to move the open-lid logic directly into Disman.