refactor(kded): remove ideal and laptop config generation
The lid closed logic was removed previously. The other case, being docked, was
not hooked up in the first place. Therefore remove all logic for generating an
"ideal" config for laptops.

What remains is the logic to adapt a config through user action on the popup.
But this is currently deactivated.