fix(kcm): snap to other outputs without gap
There was an explicit one pixel surplus added when snapping one output edge
to another one. This lead to not respected output boundaries on X11.

It is not clear why it felt necessary to add the one pixel surplus but tests
show that just removing it fixed the issue on X11 and still worked fine with
Wayland sessions of KWinFT and sway.