fix(kcm): generate pot-file relative to source root directory
In Weblate the paths to files are shown to translators for information and
they are relative to the repository root.

xgettext annotates the strings with the file path and line number in a comment.
To make the Weblate functionality work these file paths must have the same
relative root directory.

For that we change in our Messages.sh script to the source root directory and
then find all files in the respective sub-directory for xgettext.