fix(kcm): add correct spacing to the swipeview
With an spacing value = 0, the second delegate of the SwipeView
will be visible in the right margin region of the first delegate.

This shows up mainly when the FormLayout goes into narrow mode,
since then the items are left aligned and more likely to get into
the visible part.

CCBUG: 434383